requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
peewee==3.16.2
psycopg2-binary==2.9.6
PyYAML==6.0
requests==2.30.0
urllib3==2.0.2
